Move all non-browser sources to src/
In preparation for adding support for TypeScript, which will require
a build step, move all sources to a subdirectory, so relative paths
to assets and data files remain the same from the source or build
directory.
