+++ package.json
@@ -4,8 +4,8 @@
   "license": "Apache-2.0",
   "private": true,
   "scripts": {
-    "start": "node main.js",
-    "lint": "eslint ./routes ./service ./util",
+    "start": "node src/main.js",
+    "lint": "eslint ./src",
     "test": "nyc ./tests/everything.sh",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "debug": "nodemon --legacy-watch --watch ./routes --watch ./util --inspect=0.0.0.0:5858 --nolazy main.js"
