+++ src/routes/devices.js
